[Live] Fix bug where child components would sometimes fail this step
The error would be:

Failed to execute 'replaceChild' on 'Node': The node to be replaced is not a child of this node.

I'm not sure what caused that (it seemed to be related to using a non-div root child component element), but replaceWith() works great.

Also updated the outdated js built file.