[Live] Ignore problems with writing bad types for writable paths
The motivating use-case is where you have a writable property that starts
null but the setter method does NOT allow null. When that null is rehydrated later,
we attempt to call the setter with a null value, which explodes.

This makes for a better UX... though with a sprinkle of magic to make it happen.