[live-component docs] Rearrange information about using custom Stimulus controllers
In the section 'Working with the Component in JavaScript', the preamble mentions that you can create a custom Stimulus controller and attach it to or put it around the root component element in order to control its behaviour.

How to attach a custom controller is only explained further down in the subsection 'Adding a Stimulus Controller to your Component Root Element'. This subsection is currently buried beneath an intervening subsection on a different topic, 'JavaScript Component Hooks', making it easy for a reader to overlook.

As such, to improve the logical flow of the documentation, I have swapped these two subsections around.

It is especially important that this information is presented clearly because it has caused confusion in the past; see e.g. #687.